Time: 0 ms (100%), Space: 7.5 MB (50.16%) - LeetHub

+++ 0441-arranging-coins/0441-arranging-coins.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     int arrangeCoins(int n) {
-        long long temp = (8*n)+1;
-        temp = (sqrt(temp)-1)/2;
-        return temp;
+        long double temp = (8 * static_cast<long double>(n)) + 1;
+        temp = (sqrt(temp) - 1) / 2;
+        return static_cast<int>(temp);
     }
 };
